orient_requirements.txt
#Rahul's computer setup

bash <(curl -Ls http://git.io/eUx7rg)
sudo pip install mysql-python

/usr/local/mysql-5.6.22-osx10.8-x86_64/support-files/mysql.server start
mysqladmin -u root -p'WCYBVBK6' password thebakery
mysql -u root -p
    CREATE DATABASE orient;




python manage.py migrate

python manage.py createsuperuser


sudo scp -i ~/Downloads/rahulkey.pem ubuntu@ec2-54-198-19-77.compute-1.amazonaws.com:/data/visits_20150128.tsv visits_20150128.tsv

mysql -u root -p --local-infile
user orient;

CREATE TABLE `user_events` (
  `uid` varchar(100) DEFAULT NULL,
  `start_time` bigint(20) DEFAULT NULL,
  `visit_id` int(11) DEFAULT NULL,
  `etype` varchar(100) DEFAULT NULL,
  `url` varchar(2048) DEFAULT NULL,
  `is_conversion` tinyint(1) DEFAULT NULL,
  `referrer` varchar(2048) DEFAULT NULL,
  `ip_address` varchar(100) DEFAULT NULL,
  `user_agent` varchar(1000) DEFAULT NULL,
  `element` varchar(1000) DEFAULT NULL,
  `element_txt` varchar(1000) DEFAULT NULL,
  `css_class` varchar(1000) DEFAULT NULL,
  `path` varchar(1000) DEFAULT NULL,
  `log_time` bigint(20) DEFAULT NULL,
  `title` varchar(1000) DEFAULT NULL,
  `ref_domain` varchar(500) DEFAULT NULL,
  `img_src` varchar(1000) DEFAULT NULL,
  `label` varchar(1000) DEFAULT NULL,
  `start_hc` int(11) DEFAULT NULL,
  `href` varchar(1000) DEFAULT NULL
) ENGINE=InnoDB DEFAULT CHARSET=latin1;



LOAD DATA LOCAL  INFILE '/data/visits_20150130.tsv' into table user_events FIELDS terminated by '\t' lines terminated by '\n' (uid, start_time, ip_address, user_agent,referrer, visit_id, url, etype, log_time, is_conversion, element, element_txt, css_class, path, title, img_src, label, href, start_hc, ref_domain);
 



#on ubuntu server

ssh -i ~/Downloads/rahulkey.pem ubuntu@ec2-54-198-19-77.compute-1.amazonaws.com

#http://www.nickpolet.com/blog/deploying-django-on-aws/1/
sudo apt-get update
sudo apt-get install apache2 libapache2-mod-wsgi
sudo apt-get install python-pip
sudo pip install django
sudo apt-get install mysql-server python-mysqldb

sudo vim /etc/apache2/sites-enabled/orient.conf

WSGIScriptAlias / /data/orient/mysite/mysite/wsgi.py
WSGIPythonPath /data/orient/mysite 
<Directory /data/orient/mysite/mysite>
    <Files wsgi.py>
        Order deny,allow
        Require all granted
    </Files>
</Directory>

<Directory /data/orient/mysite/static>
    Require all granted
</Directory>

<Directory /data/orient/mysite/media>
    Require all granted
</Directory>

#http://stackoverflow.com/questions/1795176/how-to-change-mysql-data-directory
#http://thomas.broxrost.com/2008/08/21/persistent-django-on-amazon-ec2-and-ebs-the-easy-way/
service mysql stop
sudo cp -R -p /var/lib/mysql /data
sudo vim /etc/mysql/my.cnf 
datadir         = /data/mysql

sudo vim /etc/apparmor.d/usr.sbin.mysqld
Look for lines beginning with /var/lib/mysql. Change /var/lib/mysql in the lines with the new path.

sudo /etc/init.d/apparmor reload
sudo apt-get install git
sudo git clone https://github.com/mrbhandari/orient.git
cd orient/mysite
mysql -u root -p
    CREATE DATABASE orient;
sudo python manage.py migrate

python manage.py createsuperuser

#load data
 mysql -u root -p --local-infile
 use orient;
LOAD DATA LOCAL  INFILE '/data/all_visits.tsv' into table user_events FIELDS terminated by '\t' lines terminated by '\n' (uid, start_time, ip_address, user_agent,referrer, visit_id, url, etype, log_time, is_conversion, element, element_txt, css_class, path);
